Week 5 Dictionary (C)
Similarly to Filter, I have just uploaded the single file that contains code written by me to solve the pset (concerning writing hash function and recursively loading and freeing a dictionary as efficiently as possible to allow the spellcheck program to work). In order to run the program as intended, the Makefile, speller.c and dictionary.h files available from CS50x would also be required.